requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
textblob>=0.17.1
nltk>=3.8.0
urllib3>=2.0.0